+++ app/src/main/java/org/simple/clinic/patientattribute/entry/BMIEntryEvent.kt
@@ -18,4 +18,3 @@ data object BackPressed : BMIEntryEvent() {
   override val analyticsName = "BMI Entry::Back Pressed"
 }
 
-data object BMISaved : BMIEntryEvent()