+++ app/src/main/java/org/simple/clinic/summary/PatientSummaryEffectHandler.kt
@@ -96,6 +96,7 @@ class PatientSummaryEffectHandler @AssistedInject constructor(
         .addTransformer(MarkReminderAsShown::class.java, markReminderAsShown(schedulersProvider.io()))
         .addTransformer(LoadDataForBackClick::class.java, loadDataForBackClick(schedulersProvider.io()))
         .addTransformer(LoadDataForDoneClick::class.java, loadDataForDoneClick(schedulersProvider.io()))
+        .addTransformer(LoadBMiReading::class.java, loadBmiReading())
         .addTransformer(TriggerSync::class.java, triggerSync())
         .addTransformer(FetchHasShownMissingPhoneReminder::class.java, fetchHasShownMissingPhoneReminder(schedulersProvider.io()))
         .addTransformer(LoadMedicalOfficers::class.java, loadMedicalOfficers())
@@ -103,6 +104,7 @@ class PatientSummaryEffectHandler @AssistedInject constructor(
         .addConsumer(PatientSummaryViewEffect::class.java, viewEffectsConsumer::accept)
         .addTransformer(LoadPatientRegistrationData::class.java, checkPatientRegistrationData())
         .addTransformer(CheckIfCDSSPilotIsEnabled::class.java, checkIfCDSSPilotIsEnabled())
+        .addTransformer(LoadBMIFeature::class.java, loadBMIFeature())
         .addTransformer(LoadLatestScheduledAppointment::class.java, loadLatestScheduledAppointment())
         .addConsumer(UpdatePatientReassignmentStatus::class.java, { updatePatientReassignmentState(it.patientUuid, it.status) }, schedulersProvider.io())
         .addTransformer(CheckPatientReassignmentStatus::class.java, checkPatientReassignmentStatus())
@@ -115,6 +117,7 @@ class PatientSummaryEffectHandler @AssistedInject constructor(
         .addTransformer(SaveCVDRisk::class.java, saveCVDRisk())
         .addTransformer(LoadStatinInfo::class.java, loadStatinInfo())
         .addConsumer(UpdateTobaccoUse::class.java, { updateTobaccoUse(it.patientId, it.isSmoker, it.isUsingSmokelessTobacco) }, schedulersProvider.io())
+        .addTransformer(CreateNewBMIEntry::class.java, createNewBMIEntry())
         .build()
   }
 
@@ -388,9 +391,21 @@ class PatientSummaryEffectHandler @AssistedInject constructor(
           .map {
             val currentFacilityId = currentFacility.get().uuid
             CDSSPilotStatusChecked(isPilotEnabledForFacility =
-            country.isoCountryCode == Country.ETHIOPIA ||
-                country.isoCountryCode == Country.SRI_LANKA ||
-                cdssPilotFacilities.get().contains(currentFacilityId)
+                country.isoCountryCode == Country.ETHIOPIA ||
+                    country.isoCountryCode == Country.SRI_LANKA ||
+                    cdssPilotFacilities.get().contains(currentFacilityId)
+            )
+          }
+    }
+  }
+
+  private fun loadBMIFeature(): ObservableTransformer<LoadBMIFeature, PatientSummaryEvent> {
+    return ObservableTransformer { effects ->
+      effects
+          .observeOn(schedulersProvider.io())
+          .map {
+            BMIFeatureLoaded(
+                isEnabled = feature.isEnabled(Feature.ShowBMIContainer)
             )
           }
     }
@@ -662,4 +677,34 @@ class PatientSummaryEffectHandler @AssistedInject constructor(
           .map(::FetchedHasShownMissingPhoneReminder)
     }
   }
+
+  private fun loadBmiReading(): ObservableTransformer<LoadBMiReading, PatientSummaryEvent> {
+    return ObservableTransformer { effects ->
+      effects
+          .observeOn(schedulersProvider.io())
+          .map {
+            val patientAttribute = patientAttributeRepository.getPatientAttributeImmediate(
+                patientUuid = it.patientUuid
+            )
+            BMIReadingLoaded(patientAttribute?.bmiReading)
+          }
+    }
+  }
+
+  private fun createNewBMIEntry(): ObservableTransformer<CreateNewBMIEntry, PatientSummaryEvent> {
+    return ObservableTransformer { createNewBMIEntries ->
+      createNewBMIEntries
+          .observeOn(schedulersProvider.io())
+          .map { createNewBMIEntry ->
+            patientAttributeRepository.save(
+                bmiReading = createNewBMIEntry.reading,
+                patientUuid = createNewBMIEntry.patientUUID,
+                loggedInUserUuid = currentUser.get().uuid,
+                uuid = uuidGenerator.v4(),
+            )
+            BMISaved(createNewBMIEntry.reading)
+          }
+    }
+  }
+
 }