+++ app/src/test/java/org/simple/clinic/summary/PatientSummaryEffectHandlerTest.kt
@@ -4,6 +4,7 @@ import io.reactivex.Completable
 import io.reactivex.Observable
 import org.junit.After
 import org.junit.Test
+import org.mockito.kotlin.any
 import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.verify
@@ -90,6 +91,7 @@ class PatientSummaryEffectHandlerTest {
       remoteConfigService = NoOpRemoteConfigService(DefaultValueConfigReader()),
       overrides = mapOf(
           Feature.Screening to true,
+          Feature.ShowBMIContainer to true
       )
   )
 
@@ -1092,12 +1094,13 @@ class PatientSummaryEffectHandlerTest {
   @Test
   fun `when open BMI entry sheet view effect is received, then open the BMI entry sheet`() {
     // when
-    testCase.dispatch(OpenBMIEntrySheet(patientUuid))
+    val bmiReading = BMIReading(height = 177f, weight = 53f)
+    testCase.dispatch(OpenBMIEntrySheet(bmiReading))
 
     // then
     testCase.assertNoOutgoingEvents()
 
-    verify(uiActions).openBMIEntrySheet(patientUuid)
+    verify(uiActions).openBMIEntrySheet(bmiReading)
     verifyNoMoreInteractions(uiActions)
   }
 
@@ -1110,5 +1113,34 @@ class PatientSummaryEffectHandlerTest {
     testCase.assertNoOutgoingEvents()
     verify(uiActions).openCholesterolEntrySheet(patientUuid = patientUuid)
     verifyNoMoreInteractions(uiActions)
+  }
+
+  @Test
+  fun `when load bmi reading effect is received, then bmi reading should be loaded`() {
+    // when
+    testCase.dispatch(LoadBMiReading(patientUuid = patientUuid))
+
+    //then
+    testCase.assertOutgoingEvents(BMIReadingLoaded(null))
+  }
+
+  @Test
+  fun `when create new bmi entry effect is received, then save bmi`() {
+    // when
+    val bmiReading = BMIReading(height = 177f, weight = 53f)
+
+    testCase.dispatch(CreateNewBMIEntry(patientUuid, bmiReading))
+
+    //then
+    testCase.assertOutgoingEvents(BMISaved(bmiReading))
+  }
+
+  @Test
+  fun `when load bmi feature effect is received, then bmi feature should be loaded`() {
+    // when
+    testCase.dispatch(LoadBMIFeature)
+
+    //then
+    testCase.assertOutgoingEvents(BMIFeatureLoaded(true))
   }
 }