+++ app/src/test/java/org/simple/clinic/summary/PatientSummaryUpdateTest.kt
@@ -2692,7 +2692,21 @@ class PatientSummaryUpdateTest {
   }
 
   @Test
-  fun `when add bmi button is clicked, then open the bmi entry sheet`() {
+  fun `when BMI feature is loaded, then update the model`() {
+    val model = defaultModel
+        .patientSummaryProfileLoaded(patientSummaryProfile)
+
+    updateSpec
+        .given(model)
+        .whenEvent(BMIFeatureLoaded(true))
+        .then(assertThatNext(
+            hasModel(model.bmiVisibilityUpdated(true)),
+            hasNoEffects()
+        ))
+  }
+
+  @Test
+  fun `when add bmi button is clicked, open bmi entry sheet`() {
     val model = defaultModel
         .patientSummaryProfileLoaded(patientSummaryProfile)
 
@@ -2700,22 +2714,82 @@ class PatientSummaryUpdateTest {
         .given(model)
         .whenEvent(AddBMIClicked)
         .then(assertThatNext(
-            hasEffects(OpenBMIEntrySheet(model.patientUuid)),
+            hasEffects(OpenBMIEntrySheet(model.bmiReading)),
             hasNoModel()
         ))
   }
 
   @Test
-  fun `when BMI reading is added, then calculate the cvd risk`() {
+  fun `when bmi reading is loaded, then update the model`() {
     val model = defaultModel
         .patientSummaryProfileLoaded(patientSummaryProfile)
 
+    val bmiReading = BMIReading(
+        height = 165f,
+        weight = 60f
+    )
+
     updateSpec
         .given(model)
-        .whenEvent(BMIReadingAdded)
+        .whenEvent(BMIReadingLoaded(bmiReading))
+        .then(assertThatNext(
+            hasNoEffects(),
+            hasModel(model.bmiReadingsLoaded(bmiReading))
+        ))
+  }
+
+  @Test
+  fun `when BMI reading is added, then create new bmi entry`() {
+    val model = defaultModel
+        .patientSummaryProfileLoaded(patientSummaryProfile)
+
+    val bmiReading = BMIReading(
+        height = 165f,
+        weight = 60f
+    )
+    updateSpec
+        .given(model)
+        .whenEvent(BMIReadingAdded(bmiReading))
+        .then(assertThatNext(
+            hasEffects(CreateNewBMIEntry(patientUuid, bmiReading)),
+            hasNoModel()
+        ))
+  }
+
+  @Test
+  fun `when BMI reading is saved, then calculate non lab based cvd risk and update the model`() {
+    val model = defaultModel
+        .patientSummaryProfileLoaded(patientSummaryProfile)
+
+    val bmiReading = BMIReading(
+        height = 165f,
+        weight = 60f
+    )
+
+    updateSpec
+        .given(model)
+        .whenEvent(BMISaved(bmiReading))
         .then(assertThatNext(
             hasEffects(CalculateNonLabBasedCVDRisk(patientSummaryProfile.patient)),
-            hasNoModel()
+            hasModel(model.bmiReadingsLoaded(bmiReading))
+        ))
+  }
+
+  @Test
+  fun `when BMI reading is loaded, then update the model`() {
+    val model = defaultModel
+        .patientSummaryProfileLoaded(patientSummaryProfile)
+
+    val bmiReading = BMIReading(
+        height = 165f,
+        weight = 60f
+    )
+    updateSpec
+        .given(model)
+        .whenEvent(BMIReadingLoaded(bmiReading))
+        .then(assertThatNext(
+            hasNoEffects(),
+            hasModel(model.bmiReadingsLoaded(bmiReading))
         ))
   }