+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 ### Changes
 
 - Update Titration nudge threshold for diabetic patients in SriLanka from `140/90` to `130/80`
+- Add BMI container to patient registration screen
+- Add BMI container to patient summary screen
 
 ## 2026.05.11